+++ common/src/YM2612.c
@@ -6,6 +6,7 @@
 /***
  * TODO: check if PCM playback is also possible?
  */
+#include <stdint.h>
 #include "Enums.h"
 #include "SoundSystem.h"
 
@@ -13,4 +14,4 @@ void stopSounds(void) {}
 
 void soundTick(void) {}
 
-void playSound(int ununused) {}
+void playSound(uint8_t ununused) {}
